move schema getter to new file
