+++ firebase-vscode/src/data-connect/gql_schema_helpers.ts
